refactor(examples/web-email): inline useMailbox and useThreads hooks
